services/version_scraper/requirements.txt
flask
bs4
playwright
pymongo
jsonschema
python-dateutil